requirements-pipeline.txt
pandas==2.2.2
duckdb==1.1.3
requests==2.32.3
urllib3==2.2.2
huggingface_hub==0.24.6
google-cloud-storage==2.18.2
python-dateutil==2.9.0.post0
pyarrow